+++ services/mgmt/applicationinsights/src/package_2021_10/operations.rs
@@ -129,7 +129,7 @@ pub mod live_token {
                                 Ok(rsp_value)
                             }
                             status_code => Err(azure_core::error::Error::from(azure_core::error::ErrorKind::HttpResponse {
-                                status: status_code as u16,
+                                status: status_code,
                                 error_code: None,
                             })),
                         }
